requirements.txt
# LTS Bot backend (Python 3.9+)
wheel>=0.38.0
python-dotenv>=1.0.0
websockets>=12.0
google-generativeai>=0.8.0
pyaudio>=0.2.14
sounddevice>=0.4.6
soundfile>=0.12.1
google-genai>=1.0.0
webrtcvad>=2.0.10


# TTS: PyPI melotts 패키지는 Windows 등에서 빌드 오류가 있음. 아래 별도 설치.
# pip install git+https://github.com/myshell-ai/MeloTTS.git
# 설치 후: python -m unidic download  (필요 시)

# Optional per-platform
# pywin32  # Windows only
